requirements.txt
Ipykernel
PyYAML
dotenv
langchain_google_genai
langchain
setuptools
langchain_astradb
pandas

-e .